requirements.txt
badgrclient
PyYAML
datanommer.models
zope.sqlalchemy
requests
psutil
fedora-messaging
requests-gssapi
backoff
